Added slam command
exposed as command the implementation by Anathema: kills all
processes using the tomb (denying it to be closed) and umounts

tomb (the commandline script) is not supposed to be interactive
we leave interactivity to be implemented by wrappers as tomb-open
so it should stop to ask the used if to slam, rather than offer
it as a command. now also tomb-status has the menu option to slam.