[Optparsing] Support options in whatever position
However, it has a precondition that developers MUST respect:
  An option CAN'T have differente meanings/behaviour in different subcommands.
  For example, "-s" means "size" and accept an argument. If you are tempted to add
  an option "-s" (that means, for example "silent", and doesn't accept an argument)
This, however, make sense even from an usability point of view.
Also, option with optional argument are prohibited.

The technique is a "double scan": first, we scan options passing
_every_ possible subcommand options to zparseopts.
This way, we can distinguish between option argument and parameters.
So, we recognize which is the subcommand.
Then, parse again