tomb-open runs ignoring if swap is in use
Tomb-open is a lightweight wrapper and currently has no way to warn the user about
a swap being present. for now we simply deactivate the check.

In future, tombs will be created with a new GUI and this wrapper will be left
only for xdg registered operations (file manager right click, mime-types etc.)

This commit fixes issue #62