Security check for extra keyslots
when a tomb is opened we check if other slots besides the master (0)
are enabled and if so we warn the user that there can be a backdoor.
LUKS permits the presence of multiple keys that can open a single
volume, but this weakens the security of it and we are not really
interested in implementing it.