Fixed broken images in Tomb_User_Manual.org
There's two images in Tomb_User_Manual.org that were broken. The first
one was tomb_and_bats.png and the second was foster_privacy.png.

Image tomb_and_bats.png was relinked. The link to foster_privacy.png was
removed because I couldn't find the image in the repository.

Pull request #195 had an error with the URL of the image. This pull
request fixes that.