Man-Page: Small fixes
Mostly intended to fix the URLs at the end.
The original page for dm-crypt pointed to the google-code page of cryptsetup and the google-code page itself redirects the user to gitlab, where the project is now hosted.
So replaced the URLs with their gitlab counterpart.

Made a note about the mount behaviour which got enhanced by ca012e8

Additionally some smaller typos corrected