slam: restore implicit "all" if no tombname given
With 653609a the behaviour of slam changed slightly.
It shared the same syntax as regular close, which avoids any action if more than one tomb is opened.
Previously slam would generally close all tombs.
It is now changed insofar, that it is again possible to slam all if no argument is given.

Closes #569