Set TYPESET_SILENT
In ZSH typeset (and therefore local) will print the value of a variable
to stdout if it is already defined and not assigned a value.
This will surface in loops like the one added with 9f79e8b which adjusts
ask_password() with a better detection logic for pinentry backends.
If the loop doesn't end with the first checked backend a password
couldn't be successfully retrieved.
This could also be addressed by explicitly unsetting the
respective variables, but it is in general sensible to avoid accidently
printing values to stdout in shell scripts. Especially if the output is
being catched.

Closes #575