ask_password: update variables
Move all variable definitions to the beginning of the function.
The for loop doesn't restrict the scope of a variable (still accessable
after the loop ended) and may cause unwanted behaviour if typeset/local
is run again on an already existing variable.
Add "types" were it is sensible and remove variables which aren't used.