Add option --pinentry
This allows to specify a specifc pinentry program which is tried first
while finding a working pinentry program.
For input is only the part needed after "pinentry-".
This isn't a general option to specify something completly different as
password input (yet).

The update to the manpage includes the new option and an update to the
pinentry section. With the way how pinentry is nowadays detected, the
manual force of a specific pinentry via symlink shouldn't be encouraged.
And if a specific one should be used, one can use now the new option.