Uniform usage of print or _print() helper
As a zsh script the builtin print should be used instead of echo. And
mostly done that way. Therefore replace the few echo calls with print.

And additionally replace cat construct for printing the system utils.
This part would even appear if "tomb --version -q" would be called
(nonsensical in a way). With this only the empty print will cause
newlines. But cannot be replaced as the current wrapper _print doesn't
allow empty input.
And those cat thingies always look ugly and it offers no real benefit
in that location.