_sudo: unify branches into a singular flow
If --sudo was supplied it would behave differently than without.
More verbose, different output, but in general the same behaviour as
without.
Additionally "--sudo sudo" wasn't working.

Now this is organized in one branch instead of two. Has the benefit that
instead of a direct failure when the argument of "--sudo" isn't found,
it will then fallback to sudo and doas. Both tools were checked upon if
--sudo wasn't used and counted as defaults.

Additionally adjust the man page in that regard.