gpg_decrypt: drop GnuPG 1.4.11 workaround
Remove last remnant of 9e820f3de6a1b3ecab0662d0ae30c0b79d6d4e43
The latest release of gpg1 is 1.4.23 from 2018. It is safe to assume, that there
is no modern distribution, which ships this specific gpg1 version.
As the workaround itself only applies to 1.4.11, any version before or after seem
to be safe.
This removal doesn't remove support for gpg1 in general as it isn't end-of-life.
All the options in use are available with gpg1 and only a broken --status-fd
seemed to be of concern.

Closes #581