Switch password input for gpg operations
Formerly a combined password+key divided by a newline was piped into the
gpg operation which sets --passphrase-fd 0 to get the password from
STDIN. This gets stripped from from the pipe and only the encrypted data
remains.
This has some consequences for entered passwords in general.
Mainly \ and % related.
\ can be used for control sequences like \n, \t and \f.
If those are used in passwords various stuff can happen.
\n not at the end: Only the password up to the backslash is used and
valid.
\n at the end: <ToDo>
\t not at the end: <ToDo>
\t at the end: <ToDo>
\f
\f

<Description for % stuff>

In that regard change how gpg gets the password.
Two ways for improvements:
1) use a specific file descriptor --passphrase-fd 3 and
redirect the password via "3<<<$password".

2) switch to --passphrase-file and use an anonymous pipe <(print -R -n -
   "$password")
   In general the use of --passphrase-file is discouraged, but that
   should only apply if a real file is used. With the pipe only the
   respective file descriptor should be seen and not its content. Which
   only gpg is allowed to have access. And if gpg is completed the pipe
   and its content is gone.

This solutions uses the second way as it isn't completly sure if the
content of the redirect will be visible or not.

What changes:
\n not at the end: Assuan still outputs a formated password, but the
complete password is used.
\n at the end: <ToDo>
\t not at the end: password is functional
\t at the end: <ToDo>
\f
\f

<Does it help with % ?>