Added support for two-factor HW passkey (FIDO2)
Instead of a password, a fido2/ctap2 hardware passkey can be used to
lock/unlock tombs.

When using `--fido2` option, keys are forged as usual, but they prompt
automatically detected fido2 hardware keys when a password should be
supplied. The .key file with metadata does not contain any secrets.

The option is also used for lock/open operations. Password interactions
are replaced with FIDO2 passkey interaction (requiring a pin to unlock
the token).